RD-2002 Extract shared widget configuration and definition to common
Allows reusing the same widget configuration and definition for both
top-level and drilled-down Deployments View widgets.
