+++ widgets/common/src/deploymentsView/index.ts
@@ -1,17 +1,22 @@
-// NOTE: names prefixed with D so they can be exported inside namespace with a different name
-import * as DTable from './table';
+import * as Table from './table';
 import * as DCommon from './common';
-import DDetailsPane from './detailsPane';
-import type { Deployment, DeploymentsResponse, SharedDeploymentsViewWidgetConfiguration } from './types';
+import DetailsPane from './detailsPane';
+import type { Deployment, DeploymentsResponse } from './types';
+import { SharedDeploymentsViewWidgetConfiguration, sharedConfiguration, sharedDefinition } from './configuration';
 import './styles.scss';
 
 declare global {
     namespace Stage.Common.DeploymentsView {
-        const Table: typeof DTable;
-        const Common: typeof DCommon;
-        const DetailsPane: typeof DDetailsPane;
-        namespace Types {
-            export { Deployment, DeploymentsResponse, SharedDeploymentsViewWidgetConfiguration };
+        // NOTE: necessary rename to DCommon, since `Common` resolves to `Stage.Common`, not the `Common` import
+        export { Table, DCommon as Common, DetailsPane, sharedDefinition };
+
+        // NOTE: no-namespace rule does not detect that `export namespace` are in a `declare` context
+        /* eslint-disable @typescript-eslint/no-namespace */
+        export namespace Configuration {
+            export { SharedDeploymentsViewWidgetConfiguration, sharedConfiguration };
+        }
+        export namespace Types {
+            export { Deployment, DeploymentsResponse };
         }
     }
 }
@@ -19,9 +24,11 @@ declare global {
 Stage.defineCommon({
     name: 'DeploymentsView',
     common: {
-        DetailsPane: DDetailsPane,
+        sharedDefinition,
+        DetailsPane,
         Common: DCommon,
-        Table: DTable,
+        Table,
+        Configuration: { sharedConfiguration },
         Types: {}
     }
 });
